+++ HISTORY.rst
@@ -150,6 +150,7 @@
 - Add locking for interval sources. This reduces the chance of repeated documents from long running email downloads.
 - Add the option to enable or disable parsing when uploading a document for each document type.
 - Add a new setting option to enable automatic parsing for each new document type created.
+- Add support for HTML bodies to the user mailers.
 
 2.7.3 (2017-09-11)
 ==================